GPII-2634: Adjustments to the combined textfieldSlider and stepper widget
The model value corresponding to the slider widget is now updated only
when the user lets go of the slider thumb. While it is being dragged, the
user can still see that the value in the textfield input is updated.
However, these changes are only temporary and will not be reported to any
outer components.